refactor: fix directory structure
