+++ src/configs/plain.js
