+++ src/configs/react.js
