+++ src/configs/vue.js
